Split entries & add translations types
