+++ api/src/models/entry/index.ts
@@ -1,0 +1,11 @@
+import { t } from "elysia";
+import { Episode, MovieEntry, Special } from "../entry";
+
+export const Entry = t.Union([Episode, MovieEntry, Special]);
+export type Entry = typeof Entry.static;
+
+export * from "./episode";
+export * from "./movie-entry";
+export * from "./special";
+export * from "./extra";
+export * from "./unknown-entry";
